Flip review fixes: make entry-field helpers structurally fail-silent
Code-review follow-up (no behavior change on valid input):
- Wrap _flip_entry_fields + _flip_scan_ctx bodies in top-level try/except so
  they return {} (or partial) on ANY internal error instead of propagating
  into scan_and_trade / open_position. The fail-silent invariant is now
  enforced by construction, not by the helpers happening to be non-throwing
  (the unguarded PAIR_TREND_FILTER call site at ~7178 relied on the latter).
- Align PhantomFlip.entry_quality_score Integer→Float to match Order's column
  (avoids int/float drift in a cross-table join); migration adds it as FLOAT.

Verified: throwing input → {} (no raise); all emitted keys still valid on
both Order and PhantomFlip; quality_score now Float/Float.

+++ models.py
@@ -620,7 +620,7 @@ class PhantomFlip(Base):
     entry_bear_pct = Column(Float, nullable=True)
     entry_pair_volume_24h_usd = Column(Float, nullable=True)
     entry_pair_rank = Column(Integer, nullable=True)
-    entry_quality_score = Column(Integer, nullable=True)
+    entry_quality_score = Column(Float, nullable=True)  # Float to match Order.entry_quality_score
 
 
 class Transaction(Base):
